# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RBXL Decompiler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type DecompileSessionStatus = typeof DecompileSessionStatus[keyof typeof DecompileSessionStatus];


export const DecompileSessionStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
  error: 'error',
} as const;

export interface DecompileSession {
  id: number;
  originalName: string;
  status: DecompileSessionStatus;
  scriptCount: number;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  fileSizeBytes?: number | null;
  createdAt: string;
}

export type ScriptScriptType = typeof ScriptScriptType[keyof typeof ScriptScriptType];


export const ScriptScriptType = {
  Script: 'Script',
  LocalScript: 'LocalScript',
  ModuleScript: 'ModuleScript',
} as const;

export interface Script {
  id: number;
  sessionId: number;
  name: string;
  scriptType: ScriptScriptType;
  scriptPath: string;
  sizeBytes: number;
  isBytecode: boolean;
  createdAt: string;
}

export type ScriptWithContentScriptType = typeof ScriptWithContentScriptType[keyof typeof ScriptWithContentScriptType];


export const ScriptWithContentScriptType = {
  Script: 'Script',
  LocalScript: 'LocalScript',
  ModuleScript: 'ModuleScript',
} as const;

export interface ScriptWithContent {
  id: number;
  sessionId: number;
  name: string;
  scriptType: ScriptWithContentScriptType;
  scriptPath: string;
  sizeBytes: number;
  isBytecode: boolean;
  content: string;
  createdAt: string;
}

export type SessionStatsScriptTypes = {[key: string]: number};

export interface SessionStats {
  sessionId: number;
  totalScripts: number;
  scriptTypes: SessionStatsScriptTypes;
  totalSizeBytes: number;
  bytecodeCount: number;
  topScripts: Script[];
}

export type TopicCategory = typeof TopicCategory[keyof typeof TopicCategory];


export const TopicCategory = {
  hatch_chances: 'hatch_chances',
  egg_rates: 'egg_rates',
  pets: 'pets',
  game_mechanics: 'game_mechanics',
  currencies: 'currencies',
  shops: 'shops',
  events: 'events',
  other: 'other',
} as const;

export interface Topic {
  id: number;
  sessionId: number;
  name: string;
  category: TopicCategory;
  matchCount: number;
  createdAt: string;
}

export interface SessionAnalysis {
  sessionId: number;
  topicCount: number;
  topics: Topic[];
}

export interface TopicFinding {
  scriptName: string;
  scriptId: number;
  lineNumber: number;
  snippet: string;
  /** @nullable */
  value?: string | null;
}

export interface TopicDetail {
  id: number;
  sessionId: number;
  name: string;
  category: string;
  matchCount: number;
  findings: TopicFinding[];
  createdAt: string;
}