# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RBXL Decompiler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/scripts:
    get:
      operationId: getSessionScripts
      tags: [scripts]
      summary: List all scripts extracted from a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Script"
  /sessions/{sessionId}/scripts/{scriptId}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get full script content
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
        - name: scriptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptWithContent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Script:
      type: object
      required: [id, sessionId, name, scriptType, scriptPath, sizeBytes, isBytecode, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        name:
          type: string
        scriptType:
          type: string
          enum: [Script, LocalScript, ModuleScript]
        scriptPath:
          type: string
        sizeBytes:
          type: integer
        isBytecode:
          type: boolean
        createdAt:
          type: string
    ScriptWithContent:
      type: object
      required: [id, sessionId, name, scriptType, scriptPath, sizeBytes, isBytecode, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        name:
          type: string
        scriptType:
          type: string
          enum: [Script, LocalScript, ModuleScript]
        scriptPath:
          type: string
        sizeBytes:
          type: integer
        isBytecode:
          type: boolean
        content:
          type: string
        createdAt:
          type: string
...