# lib/api-zod/src/generated/types/script.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RBXL Decompiler API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptScriptType } from './scriptScriptType';

export interface Script {
  id: number;
  sessionId: number;
  name: string;
  scriptType: ScriptScriptType;
  scriptPath: string;
  sizeBytes: number;
  isBytecode: boolean;
  createdAt: string;
}