# lib/api-zod/src/generated/types/scriptWithContent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RBXL Decompiler API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptWithContentScriptType } from './scriptWithContentScriptType';

export interface ScriptWithContent {
  id: number;
  sessionId: number;
  name: string;
  scriptType: ScriptWithContentScriptType;
  scriptPath: string;
  sizeBytes: number;
  isBytecode: boolean;
  content: string;
  createdAt: string;
}